+++ internal/domain/audit.go
@@ -38,7 +38,7 @@ type TenantSettings struct {
 	InvoicePrefix   string    `json:"invoice_prefix"`
 	InvoiceNextSeq  int       `json:"invoice_next_seq"`
 	NetPaymentTerms int       `json:"net_payment_terms"`
-	TaxRateBP       int       `json:"tax_rate_bp"` // Basis points (1850 = 18.50%)
+	TaxRateBP       int64     `json:"tax_rate_bp"` // Basis points (1850 = 18.50%)
 	TaxName         string    `json:"tax_name,omitempty"`
 	TaxInclusive    bool      `json:"tax_inclusive"`
 	CompanyName     string    `json:"company_name,omitempty"`
